backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-generativeai>=0.3.0
python-dotenv==1.0.1
python-multipart==0.0.12
pydantic==2.9.2
pillow>=10.0.0
